Store/requirements.txt
bcrypt==4.1.3
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.1
pika==1.3.2
pluggy==1.5.0
PyJWT==2.8.0
PyMySQL==1.1.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
